chore: update workflow
